+++ src/Newbe.Mahua/NativeApi/ICoolQApi.cs
@@ -17,7 +17,7 @@
         /// <param name="QQID">目标QQ</param>
         /// <param name="msg">消息内容</param>
         /// <returns></returns>
-        int CQ_sendPrivateMsg(int AuthCode, long QQID, string msg);
+        long CQ_sendPrivateMsg(int AuthCode, long QQID, string msg);
 
         /// <summary>
         /// 发送群消息
@@ -26,7 +26,7 @@
         /// <param name="群号">目标群</param>
         /// <param name="msg">消息内容</param>
         /// <returns></returns>
-        int CQ_sendGroupMsg(int AuthCode, long 群号, string msg);
+        long CQ_sendGroupMsg(int AuthCode, long 群号, string msg);
 
         /// <summary>
         /// 发送讨论组消息
@@ -35,7 +35,7 @@
         /// <param name="讨论组号">目标讨论组</param>
         /// <param name="msg">消息内容</param>
         /// <returns></returns>
-        int CQ_sendDiscussMsg(int AuthCode, long 讨论组号, string msg);
+        long CQ_sendDiscussMsg(int AuthCode, long 讨论组号, string msg);
 
         /// <summary>
         /// 发送赞
